backend AI/requirements.txt
Flask
flask-cors
requests
cloudinary
google-generativeai
python-dotenv
edge-tts
pygame
aiofiles
